requirements.txt
opencv-python
dlib
scipy
numpy
imutils
pyqt6
rumps
matplotlib
grpcio
grpcio-tools